+++ src/XBullet.EasyTesting/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("XBullet.EasyTesting.Azure")]
+[assembly: InternalsVisibleTo("XBullet.EasyTesting.Http")]